# openapi.yaml
openapi: 3.0.3
info:
  title: WODTrackr Backend API
  version: 1.0.0
  description: OpenAPI documentation for WODTrackr exercise endpoints.
servers:
  - url: /api
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Exercise:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 120
        description:
          type: string
        category:
          type: string
          enum: [weightlifting, powerlifting, gymnastics, monostructural, accessory, mobility, other]
        equipment:
          type: string
          enum: [bodyweight, barbell, dumbbell, kettlebell, medicine_ball, box, rig, rings, rope, rower, bike, ski_erg, assault_runner, jump_rope, sled, sandbag, pegboard, other]
        primary_muscle_group:
          type: string
          maxLength: 50
        is_public:
          type: boolean
        created_by:
          type: integer
          nullable: true
        created_by_username:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - equipment
        - is_public
        - created_at
        - updated_at
    ExerciseCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [weightlifting, powerlifting, gymnastics, monostructural, accessory, mobility, other]
        equipment:
          type: string
          enum: [bodyweight, barbell, dumbbell, kettlebell, medicine_ball, box, rig, rings, rope, rower, bike, ski_erg, assault_runner, jump_rope, sled, sandbag, pegboard, other]
        primary_muscle_group:
          type: string
        is_public:
          type: boolean
      required:
        - name
    ExerciseUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [weightlifting, powerlifting, gymnastics, monostructural, accessory, mobility, other]
        equipment:
          type: string
          enum: [bodyweight, barbell, dumbbell, kettlebell, medicine_ball, box, rig, rings, rope, rower, bike, ski_erg, assault_runner, jump_rope, sled, sandbag, pegboard, other]
        primary_muscle_group:
          type: string
        is_public:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          oneOf:
            - type: string
            - type: object
paths:
  /wodtrackr/exercises/:
    get:
      summary: List exercises
      description: Returns public exercises plus the authenticated user's exercises. Supports search, filtering, and ordering.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search across name, description, and primary_muscle_group.
        - in: query
          name: category
          schema:
            type: string
            enum: [weightlifting, powerlifting, gymnastics, monostructural, accessory, mobility, other]
        - in: query
          name: equipment
          schema:
            type: string
            enum: [bodyweight, barbell, dumbbell, kettlebell, medicine_ball, box, rig, rings, rope, rower, bike, ski_erg, assault_runner, jump_rope, sled, sandbag, pegboard, other]
        - in: query
          name: muscle
          schema:
            type: string
          description: Filter by primary_muscle_group (contains).
        - in: query
          name: is_public
          schema:
            type: boolean
        - in: query
          name: mine
          schema:
            type: boolean
          description: When true, returns only exercises created by the authenticated user.
        - in: query
          name: ordering
          schema:
            type: string
            enum: [name, -name, created_at, -created_at, updated_at, -updated_at]
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Exercise'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create exercise
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Exercise'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /wodtrackr/exercises/{exercise_id}/:
    get:
      summary: Retrieve exercise
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: exercise_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Exercise'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update exercise
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: exercise_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Exercise'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete exercise
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: exercise_id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'